directory rename changes to 'encrypted Chat Protocol'
