+++ src/app/utils/encryptedChatProtocol/Packet.ts
