+++ src/app/utils/encryptedChatProtocol/commonTypes.ts
