+++ src/app/utils/encryptedChatProtocol/parser/Response.ts
