+++ src/app/utils/encryptedChatProtocol/parser/index.ts
