+++ src/app/utils/encryptedChatProtocol/responsePackets/NewToken.ts
