+++ src/app/utils/encryptedChatProtocol/responsePackets/Register.ts
