+++ src/app/utils/encryptedChatProtocol/responsePackets/ResponsePacket.ts
